+++ packages/gatsby-cli/src/reporter/loggers/ink/cli.js
@@ -3,9 +3,8 @@ import { Box, Static } from "ink"
 import isTTY from "../../../util/is-tty"
 import { trackBuildError } from "gatsby-telemetry"
 
-import Spinner from "../ink/components/spinner"
-import ProgressBar from "../ink/components/progress-bar"
-
+import { Spinner } from "../ink/components/spinner"
+import { ProgressBar } from "../ink/components/progress-bar"
 import { Message } from "../ink/components/messages"
 import Error from "./components/error"
 import Develop from "../ink/components/develop"
@@ -16,6 +15,8 @@ class CLI extends React.Component {
   state = {
     hasError: false,
   }
+
+  memoizedReactElementsForMessages = []
 
   componentDidCatch(error, info) {
     trackBuildError(`INK`, {
@@ -51,6 +52,30 @@ class CLI extends React.Component {
       )
     }
 
+    /*
+      Only operation on messages array is to push new message into it. Once
+      message is there it can't change. Because of that we can do single
+      transform from message object to react element and store it.
+      This will avoid calling React.createElement completely for every message
+      that can't change.
+    */
+    if (messages.length > this.memoizedReactElementsForMessages.length) {
+      for (
+        let index = this.memoizedReactElementsForMessages.length;
+        index < messages.length;
+        index++
+      ) {
+        const msg = messages[index]
+        this.memoizedReactElementsForMessages.push(
+          msg.level === `ERROR` ? (
+            <Error details={msg} key={index} />
+          ) : (
+            <Message key={index} {...msg} />
+          )
+        )
+      }
+    }
+
     const spinners = []
     const progressBars = []
     if (showProgress) {
@@ -71,15 +96,7 @@ class CLI extends React.Component {
     return (
       <Box flexDirection="column">
         <Box flexDirection="column">
-          <Static>
-            {messages.map((msg, index) =>
-              msg.level === `ERROR` ? (
-                <Error details={msg} key={index} />
-              ) : (
-                <Message key={index} {...msg} />
-              )
-            )}
-          </Static>
+          <Static>{this.memoizedReactElementsForMessages}</Static>
 
           {spinners.map(activity => (
             <Spinner key={activity.id} {...activity} />