+++ packages/gatsby-cli/src/util/calc-elapsed-time.ts
@@ -1,0 +1,7 @@
+import convertHrtime from "convert-hrtime"
+
+export function calcElapsedTime(startTime: [number, number]): string {
+  const elapsed = process.hrtime(startTime)
+
+  return convertHrtime(elapsed)[`seconds`].toFixed(3)
+}