+++ packages/gatsby/src/utils/create-node-id.ts
@@ -1,17 +1,17 @@
-const uuidv5 = require(`uuid/v5`)
-const report = require(`gatsby-cli/lib/reporter`)
+import uuidv5 from "uuid/v5"
+import report from "gatsby-cli/lib/reporter"
 
 const seedConstant = `638f7a53-c567-4eca-8fc1-b23efb1cfb2b`
 
 /**
  * createNodeId() Generate UUID
  *
- * @param {String} id - A string of arbitrary length
+ * @param {String | Number} id - A string of arbitrary length
  * @param {String} namespace - Namespace to use for UUID
  *
  * @return {String} - UUID
  */
-function createNodeId(id, namespace) {
+export function createNodeId(id: string | number, namespace: string): string {
   if (typeof id === `number`) {
     id = id.toString()
   } else if (typeof id !== `string`) {
@@ -22,5 +22,3 @@ function createNodeId(id, namespace) {
 
   return uuidv5(id, uuidv5(namespace, seedConstant))
 }
-
-module.exports = createNodeId