+++ www/src/components/docs-breadcrumb.js
@@ -2,8 +2,10 @@
 import { jsx } from "theme-ui"
 import React from "react"
 import Link from "./localized-link"
-import ChevronRight from "react-icons/lib/md/chevron-right"
-import ChevronLeft from "react-icons/lib/md/chevron-left"
+import {
+  MdChevronRight as ChevronRight,
+  MdChevronLeft as ChevronLeft,
+} from "react-icons/md"
 import { getItemList } from "../utils/sidebar/item-list"
 import getActiveItem from "../utils/sidebar/get-active-item"
 import getActiveItemParents from "../utils/sidebar/get-active-item-parents"