+++ packages/gatsby-page-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.42](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.41...gatsby-page-utils@0.0.42) (2020-03-09)
+
+**Note:** Version bump only for package gatsby-page-utils
+
 ## [0.0.41](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.40...gatsby-page-utils@0.0.41) (2020-03-06)
 
 **Note:** Version bump only for package gatsby-page-utils